+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/generator-simple-fe
 
+## 1.0.80
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@1.3.23
+-   @sap-ux/fiori-freestyle-writer@1.2.20
+
 ## 1.0.79
 
 ### Patch Changes